ucm: main - remove cast to pointer from integer of different size warning

+++ src/ucm/main.c
@@ -507,7 +507,7 @@ static int execute_cset(snd_ctl_t *ctl, const char *cset, unsigned int type)
 	free(value);
 	if (info2) {
 		if (info2->type == SND_CTL_ELEM_TYPE_ENUMERATED)
-			free((void *)info2->value.enumerated.names_ptr);
+			free((void *)(size_t)info2->value.enumerated.names_ptr);
 		free(info2);
 	}
 	free(info);
